+++ src/views/app/newPost.vue
@@ -1,0 +1,1 @@
+<template>new</template>
